backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-docx==0.8.11
PyPDF2==3.0.1
azure-communication-email==1.0.0
sqlalchemy==2.0.23
qdrant-client==1.7.0
sentence-transformers==2.2.2
python-dotenv==1.0.0 